Controllers have been divided and were added some small changes
