+++ AirCasting/AuthViews/UserAuthenticationSession.swift
@@ -84,102 +84,3 @@ extension UserAuthenticationSession: RequestAuthorisationService {
         return request
     }
 }
-
-protocol LogoutController {
-    func logout(onEnd: @escaping () -> Void) throws
-}
-
-final class DefaultLogoutController: LogoutController {
-    @Injected private var sessionSynchronizer: SessionSynchronizer
-    @Injected private var removeDataController: RemoveDataController
-    
-    private let responseHandler = AuthorizationHTTPResponseHandler()
-
-    func logout(onEnd: @escaping () -> Void) throws {
-        if sessionSynchronizer.syncInProgress.value {
-            var subscription: AnyCancellable?
-            subscription = sessionSynchronizer.syncInProgress.receive(on: DispatchQueue.main).sink { [weak self] value in
-                guard value == false else { return }
-                self?.removeDataController.removeData()
-                subscription?.cancel()
-                onEnd()
-            }
-            return
-        }
-        // For logout we only care about uploading sessions before we remove everything
-        sessionSynchronizer.triggerSynchronization(options: [.upload], completion: {
-            DispatchQueue.main.async { self.removeDataController.removeData(); onEnd() }
-        })
-    }
-}
-
-protocol RemoveDataController {
-    func removeData()
-}
-
-final class DefaultRemoveDataController: RemoveDataController {
-    @Injected private var microphoneManager: MicrophoneManager
-    @Injected private var deauthorizer: Deauthorizable
-    @Injected private var dataEraser: DataEraser
-
-    func removeData() {
-        Log.info("[LOGOUT] Cancelling all pending requests")
-        URLSession.shared.getAllTasks { tasks in
-            tasks.forEach { $0.cancel() }
-        }
-        if microphoneManager.isRecording {
-            Log.info("[LOGOUT] Canceling recording session")
-            microphoneManager.stopRecording()
-        }
-        Log.info("[LOGOUT] Clearing user credentials")
-        do {
-            try deauthorizer.deauthorize()
-            dataEraser.eraseAllData(completion: { [weak self] result in
-                if case let .failure(error) = result {
-                    self?.failLogout(with: error)
-                }
-            })
-        } catch {
-            failLogout(with: error)
-        }
-    }
-    
-    private func failLogout(with error: Error) {
-        assertionFailure("[LOGOUT] Failed to log out \(error)")
-    }
-}
-
-protocol DeleteAccountController {
-    func deleteAccount(completion: @escaping (Result<String, Error>) -> Void) throws
-}
-
-final class DefaultDeleteAccountController: DeleteAccountController {
-    @Injected private var authorisation: RequestAuthorisationService
-    @Injected private var urlProvider: URLProvider
-    @Injected private var client: APIClient
-    @Injected private var removeDataController: RemoveDataController
-
-    private let responseHandler = AuthorizationHTTPResponseHandler()
-    
-    func deleteAccount(completion: @escaping (Result<String, Error>) -> Void) throws {
-        let url = urlProvider.baseAppURL.appendingPathComponent("api/user.json")
-        do {
-            var request = URLRequest(url: url)
-            request.httpMethod = "DELETE"
-            request.addValue("application/json", forHTTPHeaderField: "Content-Type")
-            request.addValue("application/json", forHTTPHeaderField: "Accept")
-            try authorisation.authorise(request: &request)
-            client.requestTask(for: request) { [responseHandler] result, _ in
-                switch responseHandler.handle(result) {
-                case .success(_):
-                    self.removeDataController.removeData()
-                    completion(.success("Deleted"))
-                case .failure(let error):
-                    completion(.failure(error))
-                }
-            }
-        } catch {
-            completion(.failure(error))
-        }
-    }
-}
